requirements.txt
flask>=2.3,<3.0
pymongo[srv]>=4.3
bcrypt>=4.0
pillow>=9.4
numpy>=1.23,<2.0
tensorflow==2.17.0
